split out SwitchMaker from SkyRemote
SwitchMaker is a home assistant config generator, unrelated to the SkyRemote code
