+++ skyq/sky_remote.py
@@ -2,8 +2,6 @@ import time, math, socket, struct, time
 import requests
 import json
 
-ROOT = "/home/homeassistant/.homeassistant/"
-
 class SkyRemote:
     commands={"power": 0, "select": 1, "backup": 2, "dismiss": 2, "channelup": 6, "channeldown": 7, "interactive": 8, "sidebar": 8, "help": 9, "services": 10, "search": 10, "tvguide": 11, "home": 11, "i": 14, "text": 15,  "up": 16, "down": 17, "left": 18, "right": 19, "red": 32, "green": 33, "yellow": 34, "blue": 35, "0": 48, "1": 49, "2": 50, "3": 51, "4": 52, "5": 53, "6": 54, "7": 55, "8": 56, "9": 57, "play": 64, "pause": 65, "stop": 66, "record": 67, "fastforward": 69, "rewind": 71, "boxoffice": 240, "sky": 241}
     connectTimeout = 1000
@@ -74,62 +72,3 @@ class SkyRemote:
                 print("timeout error")
                 break
             
-
-class SwitchMaker:
-
-    def __init__(self, name, room, conf_folder):
-        self._name = name
-        self._room = room
-        ROOT = conf_folder
-        self._f= open(ROOT + "skyq" + self._room.replace(" ","") + ".yaml","w+")
-        self._f.write("    skyq_pause" + self._room.replace(" ","").lower() + ":\n" + 
-                "      value_template: '{{\"off\"}}'\n" + 
-                "      friendly_name: 'pause in the " + self._room + "'\n" + 
-                "      turn_on:\n" + 
-                "        service: media_player.media_pause\n" + 
-                "        data:\n" + 
-                "          entity_id: media_player." + self._name.replace(" ","_").lower() + "\n" +
-                "      turn_off:\n" + 
-                "        service: script.placeholder\n")
-        self._f.write("    skyq_play" + self._room.replace(" ","").lower() + ":\n" + 
-                "      value_template: '{{\"off\"}}'\n" + 
-                "      friendly_name: 'play in the " + self._room + "'\n" + 
-                "      turn_on:\n" + 
-                "        service: media_player.media_play\n" + 
-                "        data:\n" + 
-                "          entity_id: media_player." + self._name.replace(" ","_").lower() + "\n" +
-                "      turn_off:\n" + 
-                "        service: script.placeholder\n")
-        self._f.write("    skyq_ff" + self._room.replace(" ","").lower() + ":\n" + 
-                "      value_template: '{{\"off\"}}'\n" + 
-                "      friendly_name: 'fastforward in the " + self._room + "'\n" + 
-                "      turn_on:\n" + 
-                "        service: media_player.media_next_track\n" + 
-                "        data:\n" + 
-                "          entity_id: media_player." + self._name.replace(" ","_").lower() + "\n" +
-                "      turn_off:\n" + 
-                "        service: script.placeholder\n")
-        self._f.write("    skyq_rw" + self._room.replace(" ","").lower() + ":\n" + 
-                "      value_template: '{{\"off\"}}'\n" + 
-                "      friendly_name: 'rewind in the " + self._room + "'\n" + 
-                "      turn_on:\n" + 
-                "        service: media_player.media_previous_track\n" + 
-                "        data:\n" + 
-                "          entity_id: media_player." + self._name.replace(" ","_").lower() + "\n" +
-                "      turn_off:\n" + 
-                "        service: script.placeholder\n")
-        
-    def addChannel(self, channel):
-        self._f.write("    skyq_" + channel.replace(" ","").lower() + self._room.replace(" ","").lower() + ":\n" + 
-                "      value_template: '{{\"off\"}}'\n" + 
-                "      friendly_name: '"+ channel + " in the " + self._room + "'\n" + 
-                "      turn_on:\n" + 
-                "        service: media_player.select_source\n" + 
-                "        data:\n" + 
-                "          entity_id: media_player." + self._name.replace(" ","_").lower() + "\n" +
-                "          source: '" + channel + "'\n" +  
-                "      turn_off:\n" + 
-                "        service: script.placeholder\n")
-    def closeFile(self):
-        self._f.close()
-
